requirements.txt
fastapi
uvicorn
aquosRemote
python-dotenv
sharp_aquos_rc @ git+https://github.com/tunatuna1733/sharp_aquos_rc.git@0a6fdda
pyyaml